requirements.txt
dvc==3.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
joblib==1.3.2